refactor(OpenGL): removing subfolders inside component, resource & utils
