+++ src/plugin/opengl/src/OpenGL.hpp
@@ -1,23 +1,23 @@
 #pragma once
 
-#include "component/FontHandle/FontHandle.hpp"
-#include "component/MaterialHandle/MaterialHandle.hpp"
-#include "component/ModelHandle/ModelHandle.hpp"
-#include "component/ShaderHandle/ShaderHandle.hpp"
-#include "component/Sprite/Sprite.hpp"
-#include "component/SpriteHandle/SpriteHandle.hpp"
-#include "component/Text/Text.hpp"
-#include "component/TextHandle/TextHandle.hpp"
-#include "component/TextureHandle/TextureHandle.hpp"
+#include "component/FontHandle.hpp"
+#include "component/MaterialHandle.hpp"
+#include "component/ModelHandle.hpp"
+#include "component/ShaderHandle.hpp"
+#include "component/Sprite.hpp"
+#include "component/SpriteHandle.hpp"
+#include "component/Text.hpp"
+#include "component/TextHandle.hpp"
+#include "component/TextureHandle.hpp"
 
 #include "plugin/PluginOpenGL.hpp"
 
-#include "resource/FontManager/FontManager.hpp"
-#include "resource/GLMeshBufferManager/GLMeshBufferManager.hpp"
-#include "resource/GLTextBufferManager/GLTextBufferManager.hpp"
-#include "resource/MaterialCache/MaterialCache.hpp"
-#include "resource/ShaderManager/ShaderManager.hpp"
-#include "resource/TextureManager/TextureManager.hpp"
+#include "resource/FontManager.hpp"
+#include "resource/GLMeshBufferManager.hpp"
+#include "resource/GLTextBufferManager.hpp"
+#include "resource/MaterialCache.hpp"
+#include "resource/ShaderManager.hpp"
+#include "resource/TextureManager.hpp"
 
 #include "system/BufferSystems.hpp"
 #include "system/ManagerSystems.hpp"
@@ -25,11 +25,11 @@
 #include "system/ShaderSystems.hpp"
 #include "system/WindowSystems.hpp"
 
-#include "utils/Font/Font.hpp"
-#include "utils/GLMeshBuffer/GLMeshBuffer.hpp"
-#include "utils/GLTextBuffer/GLTextBuffer.hpp"
-#include "utils/Loader/Loader.hpp"
-#include "utils/Material/Material.hpp"
-#include "utils/MouseDragging/MouseDragging.hpp"
-#include "utils/Texture/Texture.hpp"
-#include "utils/Viewer/Viewer.hpp"
+#include "utils/Font.hpp"
+#include "utils/GLMeshBuffer.hpp"
+#include "utils/GLTextBuffer.hpp"
+#include "utils/Loader.hpp"
+#include "utils/Material.hpp"
+#include "utils/MouseDragging.hpp"
+#include "utils/Texture.hpp"
+#include "utils/Viewer.hpp"
