+++ src/plugin/opengl/src/component/Text.hpp
