+++ src/plugin/opengl/src/component/TextHandle.hpp
