+++ src/plugin/opengl/src/resource/Camera.hpp
@@ -3,7 +3,8 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 
-#include "utils/Viewer/Viewer.hpp"
+#include "utils/Viewer.hpp"
+
 namespace ES::Plugin::OpenGL::Resource {
 class Camera {
   public:
