+++ src/plugin/opengl/src/resource/GLSpriteBufferManager.hpp
