+++ src/plugin/opengl/src/resource/MaterialCache.hpp
