+++ src/plugin/opengl/src/utils/GLTextBuffer.cpp
