+++ src/plugin/opengl/xmake.lua
@@ -30,33 +30,9 @@ target("PluginOpenGL")
 
     add_includedirs("src/", {public = true})
     add_includedirs("src/exception/", {public = true})
-    add_includedirs("src/utils/Font/", {public = true})
-    add_includedirs("src/utils/Texture/", {public = true})
-    add_includedirs("src/utils/GLMeshBuffer/", {public = true})
-    add_includedirs("src/utils/GLTextBuffer/", {public = true})
-    add_includedirs("src/utils/GLSpriteBuffer/", {public = true})
-    add_includedirs("src/utils/Material/", {public = true})
-    add_includedirs("src/utils/MouseDragging/", {public = true})
-    add_includedirs("src/utils/Viewer/", {public = true})
-    add_includedirs("src/utils/Loader/", {public = true})
-    add_includedirs("src/utils/Light", {public = true})
-    add_includedirs("src/component/FontHandle/", {public = true})
-    add_includedirs("src/component/ShaderHandle/", {public = true})
-    add_includedirs("src/component/MaterialHandle/", {public = true})
-    add_includedirs("src/component/ModelHandle/", {public = true})
-    add_includedirs("src/component/TextHandle/", {public = true})
-    add_includedirs("src/component/TextureHandle/", {public = true})
-    add_includedirs("src/component/SpriteHandle/", {public = true})
-    add_includedirs("src/component/Sprite/", {public = true})
-    add_includedirs("src/component/Text/", {public = true})
-    add_includedirs("src/resource/TextureManager/", {public = true})
-    add_includedirs("src/resource/FontManager/", {public = true})
-    add_includedirs("src/resource/ShaderManager/", {public = true})
-    add_includedirs("src/resource/GLMeshBufferManager/", {public = true})
-    add_includedirs("src/resource/GLTextBufferManager/", {public = true})
-    add_includedirs("src/resource/GLSpriteBufferManager/", {public = true})
-    add_includedirs("src/resource/MaterialCache/", {public = true})
-    add_includedirs("src/resource/Camera/", {public = true})
+    add_includedirs("src/utils/", {public = true})
+    add_includedirs("src/component/", {public = true})
+    add_includedirs("src/resource/", {public = true})
     add_includedirs("src/system/", {public = true})
     add_includedirs("src/plugin/", {public = true})
     add_headerfiles("src/**.hpp")
